17e344c COMP: Removed warnings

+++ VolumeRendering/vtkOpenGLProjectedTetrahedraMapper.cxx
@@ -96,9 +96,9 @@ vtkOpenGLProjectedTetrahedraMapper::vtkOpenGLProjectedTetrahedraMapper()
   this->FloatingPointFrameBufferResourcesAllocated = false;
 
   this->Internals = new vtkOpenGLProjectedTetrahedraMapper::vtkInternals;
-  this->Internals->FrameBufferObjectId = -1;
+  this->Internals->FrameBufferObjectId = static_cast<GLuint>(-1);
   this->Internals->RenderBufferObjectIds[0]
-      = this->Internals->RenderBufferObjectIds[1] = -1;
+    = this->Internals->RenderBufferObjectIds[1] = static_cast<GLuint>(-1);
   this->Internals->OpacityTexture = 0;
 
   this->UseFloatingPointFrameBuffer = true;
@@ -233,10 +233,10 @@ void vtkOpenGLProjectedTetrahedraMapper::ReleaseGraphicsResources(vtkWindow *win
       this->FloatingPointFrameBufferResourcesAllocated = false;
 
       vtkgl::DeleteFramebuffers(1, &this->Internals->FrameBufferObjectId);
-      this->Internals->FrameBufferObjectId = -1;
+      this->Internals->FrameBufferObjectId = static_cast<GLuint>(-1);
 
       vtkgl::DeleteFramebuffers(2, this->Internals->RenderBufferObjectIds);
-      this->Internals->RenderBufferObjectIds[0] = -1;
+      this->Internals->RenderBufferObjectIds[0] = static_cast<GLuint>(-1);
     }
 
   this->Superclass::ReleaseGraphicsResources(win);
